requirements.txt
speechbrain
torchaudio
openai-whisper
fastapi
uvicorn